Update introduction.md (general questions on pronouns)
Hi Amit

Noticed that we often use the pronoun us/we/our. 

a) Do we want to keep that? Becasue for me it feels a bit strange when we try to provide a impartial source of information and then repeatedly say we/us/our. Are trying to be a blog or a wiki? 

b) If we use “we/us/our”, we should be consistent and use the form “our CTO” and then brakets or something else to show who you are.