Update introduction.md (Trigger warning!)
Hi Amit 

a) I know this is probably not how we best approach this. I think we need to agree on a way how I can make multiple changes and add general comments to the text, without creating a mess for you. 

b) I hope you don't mind me challenging your expertise in this section. I just try to approach it as a novice (i am) and give feedback on what I understand and what I dont follow 

c)  In some sections the information density is just too high. 

d) in some section it feels like your using to colloquial of a language.  (example: Changing to question mark, the face correctly become questioning (even though the SignWriting is not perfect).